+++ ChangeLog.md
@@ -18,6 +18,7 @@ This project's release branch is `master`. This log is written from the perspect
   version. As a result, you no longer need to have ghcid installed to
   use `ob run`, as we provide one for you.
 * `ob` commands now complain less on systems with umasks other than `0022`.
+* Ignore package environment files in `ob run` and `ob repl`
 
 ## v0.2.0.0 - 2019-8-17